Conflicts:
	Questor.Modules/Caching/Cache.cs
	Questor.Modules/Lookup/Settings.cs

+++ Questor.Modules/Actions/AgentInteraction.cs
@@ -90,7 +90,7 @@ namespace Questor.Modules.Actions
             //
             if (DateTime.UtcNow < _agentStandingsCheckTimeOut)
             {
-                if (((int)Cache.Instance.StandingUsedToAccessAgent == (int)0.00) && (AgentId == Cache.Instance.AgentId))
+                if (((int)Cache.Instance.StandingUsedToAccessAgent == (float)0.00) && (AgentId == Cache.Instance.AgentId))
                 {
                     if (!_agentStandingsCheckFlag)
                     {
